Rename 'iventory' to 'inventory'.
We're no longer restricted by 8.3 filenames.  This additionally fixes a
typo in a comment that referred to 'iventory' (found via the same grep).
