+++ src/global.c
@@ -40,7 +40,7 @@
 #include "includes/larnfunc.h"
 #include "includes/display.h"
 #include "includes/global.h"
-#include "includes/iventory.h"
+#include "includes/inventory.h"
 #include "includes/io.h"
 #include "includes/monster.h"
 #include "includes/scores.h"
