+++ src/monster.c
@@ -50,7 +50,7 @@
 #include <stdio.h>
 #include "includes/display.h"
 #include "includes/global.h"
-#include "includes/iventory.h"
+#include "includes/inventory.h"
 #include "includes/io.h"
 #include "includes/monster.h"
 #include "includes/spells.h"
