+++ src/moreobj.c
@@ -29,7 +29,7 @@ pray_at_altar()
 #include "includes/create.h"
 #include "includes/display.h"
 #include "includes/global.h"
-#include "includes/iventory.h"
+#include "includes/inventory.h"
 #include "includes/io.h"
 #include "includes/moreobj.h"
 #include "includes/object.h"
